Use vtable and state on GoApi to enable callbacks
